+++ src/AvroConvert/AvroObjectServices/Read/Resolvers/Double.cs
@@ -25,35 +25,7 @@ namespace SolTechnology.Avro.AvroObjectServices.Read
         {
             double value = reader.ReadDouble();
 
-            if (readType != typeof(double))
-            {
-                if (readType == typeof(int) || readType == typeof(int?))
-                {
-                    return Convert.ToInt32(value);
-                }
-
-                if (readType == typeof(short) || readType == typeof(short?))
-                {
-                    return Convert.ToInt16(value);
-                }
-
-                if (readType == typeof(decimal) || readType == typeof(decimal?))
-                {
-                    return Convert.ToDecimal(value);
-                }
-
-                if (readType == typeof(long) || readType == typeof(long?))
-                {
-                    return Convert.ToInt64(value);
-                }
-
-                if (readType == typeof(float) || readType == typeof(float?))
-                {
-                    return Convert.ToSingle(value);
-                }
-            }
-
-            return value;
+            return readType == typeof(double) ? value : ConvertValue(readType, value);
         }
     }
 }