+++ src/AvroConvert/AvroObjectServices/Read/Resolvers/Float.cs
@@ -25,35 +25,7 @@ namespace SolTechnology.Avro.AvroObjectServices.Read
         {
             float value = reader.ReadFloat();
 
-            if (readType != typeof(float) )
-            {
-                if (readType == typeof(int) || readType == typeof(int?))
-                {
-                    return Convert.ToInt32(value);
-                }
-
-                if (readType == typeof(short) || readType == typeof(short?))
-                {
-                    return Convert.ToInt16(value);
-                }
-
-                if (readType == typeof(decimal) || readType == typeof(decimal?))
-                {
-                    return Convert.ToDecimal(value);
-                }
-
-                if (readType == typeof(long) || readType == typeof(long?))
-                {
-                    return Convert.ToInt64(value);
-                }
-
-                if (readType == typeof(double) || readType == typeof(double?))
-                {
-                    return Convert.ToDouble(value);
-                }
-            }
-
-            return value;
+            return readType == typeof(float) ? value : ConvertValue(readType, value);
         }
     }
 }