+++ src/AvroConvert/AvroObjectServices/Read/Resolvers/Record.cs
@@ -18,11 +18,9 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Reflection;
 using System.Runtime.Serialization;
 using FastMember;
 using SolTechnology.Avro.AvroObjectServices.Schemas;
-using SolTechnology.Avro.AvroObjectServices.Schemas.AvroTypes;
 
 namespace SolTechnology.Avro.AvroObjectServices.Read
 {