+++ src/AvroConvert/AvroObjectServices/Write/Resolvers/Record.cs
@@ -88,7 +88,7 @@ namespace SolTechnology.Avro.AvroObjectServices.Write.Resolvers
             var calHash = Expression.Assign(nameHash,
                 Expression.Call(memberName, typeof(object).GetMethod("GetHashCode")));
             var cases = new List<SwitchCase>();
-            foreach (var propertyInfo in type.GetProperties())
+            foreach (var propertyInfo in type.GetProperties(BindingFlags.Instance | BindingFlags.Public | BindingFlags.IgnoreCase | BindingFlags.FlattenHierarchy))
             {
                 var property = Expression.Property(Expression.Convert(instance, type), propertyInfo.Name);
                 var propertyHash = Expression.Constant(propertyInfo.Name.GetHashCode(), typeof(int));